[Refactor] userInfo 컴포넌트 분리
